refactor: Update CSRF handling in calendar API requests and restructure event interfaces
- Disabled CSRF token requirement in listCalendarMembers and listSosoPointHistory functions.
- Moved event-related TypeScript interfaces from eventAPI.ts to interfaces.ts for better organization and clarity.
- Updated useEventDetailDialog to import EventDetail from the new location.
